+++ CHANGELOG.rst
@@ -22,13 +22,19 @@ Each individual change should have a link to the pull request after the descript
 Changed
 ^^^^^^^
 
+- converted OneHotEncodingTransformer to narwhals `#355 <https://github.com/lvgig/tubular/issues/355>_`
 - updated WeightsColumnMixin to use new narwhals 'is_finite' method
 - narwhalified ModeImputer `#321 <https://github.com/lvgig/tubular/issues/321>_`
 - fixed issues with all null and nullable-bool column handling in dataframe_init_dispatch
 - added NaN error handling to WeightColumnMixin
+- narwhalified BaseNumericTransformer `#358 <https://github.com/lvgig/tubular/issues/358>_`
+- narwhalified BaseCappingTransformer `#357 <https://github.com/lvgig/tubular/issues/357>_`
+- narwhalified CappingTransformer `#361 <https://github.com/lvgig/tubular/issues/361>_`
+- narwhalified OutOfRangeNullTransformer `#362 <https://github.com/lvgig/tubular/issues/362>_`
 - narwhalified MeanImputer `#344 https://github.com/lvgig/tubular/issues/344_`
 - narwhalified BaseGenericDateTransformer. As part of this updated test data handling of date columns
   across repo `#365 <https://github.com/lvgig/tubular/issues/365>_`
+- narwhalified BaseNumericTransformer `#358 https://github.com/lvgig/tubular/issues/358`
 - narwhalified DropOriginalMixin `#352 <https://github.com/lvgig/tubular/issues/352>_`
 - narwhalified BaseMappingTransformer `#367 <https://github.com/lvgig/tubular/issues/367>_`
 - placeholder